+++ api/controllers/user.go
@@ -20,6 +20,7 @@ package controllers
 
 import (
 	"context"
+
 	"github.com/e154/smart-home/api/stub/api"
 	m "github.com/e154/smart-home/models"
 	"github.com/go-playground/validator/v10"
@@ -111,7 +112,7 @@ func (c ControllerUser) GetUserList(ctx context.Context, req *api.GetUserListReq
 		return nil, c.error(ctx, nil, err)
 	}
 
-	return c.dto.User.ToListResult(items, uint32(total), req.Limit, req.Offset), nil
+	return c.dto.User.ToListResult(items, uint64(total), req.Limit, req.Offset), nil
 }
 
 // DeleteUserById ...
