+++ endpoint/role.go
@@ -20,6 +20,7 @@ package endpoint
 
 import (
 	"context"
+
 	"github.com/e154/smart-home/common"
 	m "github.com/e154/smart-home/models"
 	"github.com/e154/smart-home/system/access_list"
@@ -155,10 +156,10 @@ func (n *RoleEndpoint) Delete(ctx context.Context, name string) (err error) {
 }
 
 // Search ...
-func (n *RoleEndpoint) Search(ctx context.Context, query string, limit, offset int) (result []*m.Role, total int64, err error) {
+func (n *RoleEndpoint) Search(ctx context.Context, query string, limit, offset int64) (result []*m.Role, total int64, err error) {
 
 	if limit == 0 {
-		limit = int(common.DefaultPageSize)
+		limit = common.DefaultPageSize
 	}
 
 	result, total, err = n.adaptors.Role.Search(query, limit, offset)
